+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 ### Changed
 - Improve error handling in custom yaml unmarshalling.
+- Exit for unexpected errors occuring during when clause validation.
 
 ## 0.1.1 (2017-09-16)
 ### Changed